added calculation test and organize model test files
